Enable compiling with wasm (but compile basically nothing)
