+++ .github/workflows/test-all.yaml
@@ -30,7 +30,7 @@ jobs:
           - os: windows-latest
           - os: macos-latest
           - os: ubuntu-latest
-            target_option: "--exclude=prqlc --target=wasm32-unknown-unknown"
+            target_option: "--target=wasm32-unknown-unknown"
     steps:
       - name: 📂 Checkout code
         uses: actions/checkout@v3
