+++ prql-compiler/prqlc/Cargo.toml
@@ -1,4 +1,5 @@
 [package]
+description = "CLI interface for the PRQL compiler"
 name = "prqlc"
 
 edition.workspace = true
@@ -7,7 +8,7 @@ repository.workspace = true
 rust-version.workspace = true
 version.workspace = true
 
-[dependencies]
+[target.'cfg(not(target_family="wasm"))'.dependencies]
 anyhow = {version = "1.0.57"}
 ariadne = "0.1.5"
 atty = "0.2.14"
@@ -16,9 +17,9 @@ clio = {version = "0.2.4", features = ['clap-parse']}
 color-eyre = "0.6.1"
 env_logger = {version = "0.9.1", features = ["termcolor"]}
 itertools = "0.10.3"
-prql-compiler = { path = '..' }
+prql-compiler = {path = '..'}
 serde_json = "1.0.81"
 serde_yaml = "0.9.1"
 
-[dev-dependencies]
+[target.'cfg(not(target_family="wasm"))'.dev-dependencies]
 insta = {version = "1.26", features = ["colors", "glob", "yaml"]}
